some file structure changes
